# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SBG Internal Portal API",
    "description": "REST API for the AWS Student Builder Group internal operations portal — members, tasks, submissions, link shortener, audit logs, and analytics. All endpoints (except auth) require a valid `sbg_session` HttpOnly cookie obtained via the OTP login flow.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Relative to deployment origin" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sbg_session"
      },
      "cronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cron-secret",
        "description": "Shared secret configured via the CRON_SECRET env var. An `Authorization: Bearer <secret>` header is also accepted."
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": ["SBG_LEADER", "SECRETARY", "DIRECTOR", "MANAGER", "ASSOCIATE", "BUILDER"]
      },
      "Domain": {
        "type": "string",
        "enum": ["Technical", "Corporate", "Creatives", "General"]
      },
      "Subdomain": {
        "type": "string",
        "enum": ["Software Development", "AI & Machine Learning", "Cloud & DevOps", "Events & Operations", "Sponsorship & Finance", "HR & Admin", "PR & Marketing", "Digital Design", "Media Production"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "taskId": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date-time" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "assignmentType": { "type": "string", "enum": ["INDIVIDUAL", "DOMAIN", "SUBDOMAIN", "GENERAL"] },
          "assignedToId": { "type": "string", "nullable": true },
          "assignedToName": { "type": "string" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true },
          "createdBy": { "type": "string" },
          "createdByName": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["OPEN", "CLOSED"], "description": "Lazily flipped OPEN->CLOSED on read once the deadline has passed — no scheduled job required." },
          "totalSubmissions": { "type": "integer" },
          "reminderSentAt": { "type": "string", "format": "date-time", "nullable": true, "description": "Set once a deadline-reminder email batch has been sent for this task; cleared if the deadline is changed." }
        }
      },
      "TaskPriority": {
        "type": "string",
        "enum": ["LOW", "MEDIUM", "HIGH"]
      },
...
      "TaskUpdateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date-time", "description": "Changing this clears reminderSentAt so a new reminder cycle can fire" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "status": { "type": "string", "enum": ["OPEN", "CLOSED"] }
        }
      },
      "TaskDetail": {
        "type": "object",
        "properties": {
          "task": { "$ref": "#/components/schemas/Task" },
          "submissions": { "type": "array", "items": { "$ref": "#/components/schemas/Submission" } },
          "mySubmission": { "allOf": [{ "$ref": "#/components/schemas/Submission" }], "nullable": true },
          "canReview": { "type": "boolean" },
          "canSubmit": { "type": "boolean" },
          "canDelete": { "type": "boolean" },
          "canClose": { "type": "boolean" }
        }
      },
      "Submission": {
        "type": "object",
        "properties": {
          "submissionId": { "type": "string" },
          "taskId": { "type": "string" },
          "taskTitle": { "type": "string" },
          "memberId": { "type": "string" },
          "memberName": { "type": "string" },
          "content": { "type": "string" },
          "links": { "type": "array", "items": { "type": "string", "format": "uri" } },
          "submittedAt": { "type": "string", "format": "date-time" },
          "reviewStatus": { "type": "string", "enum": ["PENDING", "APPROVED", "REJECTED"] },
          "reviewedBy": { "type": "string", "nullable": true },
          "reviewedByName": { "type": "string", "nullable": true },
          "reviewedAt": { "type": "string", "format": "date-time", "nullable": true },
          "reviewFeedback": { "type": "string", "nullable": true, "description": "Optional comment left by the reviewer on APPROVE/REJECT" },
          "ratingAwarded": { "type": "number", "nullable": true },
          "deadline": { "type": "string", "format": "date-time" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true }
        }
      },
...
      "LeaderboardEntry": {
        "type": "object",
        "properties": {
          "memberId": { "type": "string" },
          "name": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true },
          "totalStars": { "type": "number" },
          "approvedCount": { "type": "integer" },
          "rejectedCount": { "type": "integer" },
          "pendingCount": { "type": "integer" }
        }
      },
      "AuditLog": {
        "type": "object",
        "properties": {
          "logId": { "type": "string" },
          "action": { "type": "string" },
          "performedBy": { "type": "string" },
          "performedByName": { "type": "string" },
          "targetType": { "type": "string" },
          "targetId": { "type": "string" },
          "details": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "AnalyticsResponse": {
        "type": "object",
        "properties": {
          "overview": {
            "type": "object",
            "properties": {
              "totalMembers": { "type": "integer" },
              "totalTasks": { "type": "integer" },
              "openTasks": { "type": "integer" },
              "totalSubmissions": { "type": "integer" },
              "approvalRate": { "type": "integer" }
            }
          },
          "domainStats": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "domain": { "type": "string" },
                "members": { "type": "integer" },
                "tasks": { "type": "integer" },
                "submissions": { "type": "integer" }
              }
            }
          },
          "submissionStats": {
            "type": "object",
            "properties": {
              "total": { "type": "integer" },
              "approved": { "type": "integer" },
              "rejected": { "type": "integer" },
              "pending": { "type": "integer" }
            }
          },
          "roleStats": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": { "$ref": "#/components/schemas/Role" },
                "count": { "type": "integer" }
              }
            }
          },
          "taskTrend": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string" },
                "count": { "type": "integer" }
              }
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "No valid session cookie present",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "Unauthorized" } } }
      },
      "Forbidden": {
        "description": "Authenticated but not permitted to perform this action",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "Forbidden" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "sessionCookie": [] }],
  "paths": {
...
    "/tasks/{taskId}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get task detail with submissions and caller permissions",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Task detail", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/TaskDetail" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Task not found or not visible to caller" }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update a task (title, description, deadline, status)",
        "description": "Only the task creator or presidium may edit. Reopening a CLOSED task requires reviewer privileges (presidium/director/manager).",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskUpdateRequest" } } } },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Invalid status value, or no valid fields to update" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task and cascade-delete its submissions",
        "description": "Only the task creator or presidium may delete.",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/leaderboard": {
      "get": {
        "tags": ["Leaderboard"],
        "summary": "Get the star-rating leaderboard",
        "description": "Presidium roles (SBG_LEADER, SECRETARY) are excluded — they don't participate in the rating system.",
        "parameters": [
          { "name": "domain", "in": "query", "schema": { "$ref": "#/components/schemas/Domain" } },
          { "name": "subdomain", "in": "query", "schema": { "$ref": "#/components/schemas/Subdomain" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Leaderboard", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/LeaderboardEntry" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/analytics": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get org-wide analytics overview",
        "description": "Presidium or DIRECTOR only.",
        "responses": {
          "200": { "description": "Analytics", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/AnalyticsResponse" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/audit-logs": {
      "get": {
        "tags": ["Audit Logs"],
        "summary": "List audit log entries",
        "description": "Presidium only.",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "action", "in": "query", "schema": { "type": "string" } },
          { "name": "performedBy", "in": "query", "schema": { "type": "string" }, "description": "memberId of the actor" }
        ],
        "responses": {
          "200": { "description": "Audit log list", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/AuditLog" } } } } } } },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/cron/task-reminders": {
      "get": {
        "tags": ["Cron"],
        "summary": "Email members who haven't submitted a task nearing its deadline",
        "description": "Not user-facing — meant to be invoked by an external scheduler (e.g. the included GitHub Actions workflow, EventBridge Scheduler, cron-job.org) on a recurring basis. For each OPEN task whose deadline falls within `hoursBefore` hours and that hasn't already had a reminder sent, emails every eligible member who has not yet submitted, then marks the task's `reminderSentAt`.",
        "security": [{ "cronSecret": [] }],
        "parameters": [
          { "name": "hoursBefore", "in": "query", "schema": { "type": "integer", "default": 24 }, "description": "Only consider tasks whose deadline is within this many hours from now" }
        ],
        "responses": {
          "200": {
            "description": "Reminder batch processed",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "tasksProcessed": { "type": "integer" }, "emailsSent": { "type": "integer" } } } } }
          },
          "401": { "description": "Missing or invalid cron secret" }
        }
      }
    }
  }
}